test fetch_node and friends

+++ t/21storage-dbi.t
@@ -3,6 +3,8 @@ use warnings;
 use Test::More;
 use Algorithm::SpatialIndex;
 
+my $do_unlink = !$ENV{PERL_ASI_TESTING_PRESERVE};
+
 my $tlibpath;
 BEGIN {
   $tlibpath = -d "t" ? "t/lib" : "lib";
@@ -12,22 +14,22 @@ use lib $tlibpath;
 if (not eval {require DBI; require DBD::SQLite; 1;}) {
   plan skip_all => 'These tests require DBI and DBD::SQLite';
 }
-plan tests => 3;
+plan tests => 12;
 
-my $dbfile = '21storage-dbi.test.db';
+my $dbfile = '21storage-dbi.test.sqlite';
 unlink $dbfile if -f $dbfile;
 
 my $dbh = DBI->connect("dbi:SQLite:dbname=$dbfile", "", "");
 ok(defined($dbh), 'got dbh');
 
 END {
-  unlink $dbfile;
+  unlink $dbfile if $do_unlink;
 }
 
 my $index = Algorithm::SpatialIndex->new(
   strategy => 'Test',
   storage  => 'DBI',
-  dbh_rw => $dbh,
+  dbh_rw   => $dbh,
 );
 
 isa_ok($index, 'Algorithm::SpatialIndex');
@@ -35,10 +37,29 @@ isa_ok($index, 'Algorithm::SpatialIndex');
 my $storage = $index->storage;
 isa_ok($storage, 'Algorithm::SpatialIndex::Storage::DBI');
 
-=pod
+is($storage->get_option('no_of_subnodes'), '4');
+$storage->set_option('no_of_subnodes', 5);
+is($storage->get_option('no_of_subnodes'), '5');
+my $prefix = $storage->table_prefix;
+my $res = $storage->dbh_ro->selectall_arrayref(
+  qq(SELECT id, value FROM ${prefix}_options WHERE id=?), {}, 'no_of_subnodes'
+);
+is_deeply($res, [['no_of_subnodes' => '5']], 'set_options writes to db');
 
 ok(!defined($storage->fetch_node(0)), 'No nodes to start with');
 ok(!defined($storage->fetch_node(1)), 'No nodes to start with');
+
+$storage->dbh_rw->do(
+  qq#INSERT INTO ${prefix}_nodes VALUES (0, 1., 2., 3., 4., 9, NULL, NULL, NULL)#
+);
+
+my $n = $storage->fetch_node(0);
+isa_ok($n, 'Algorithm::SpatialIndex::Node');
+is($n->id, 0, 'node id okay (manual insertion)');
+is_deeply($n->coords, [1.,2.,3.,4.], 'node coords okay (manual insertion)');
+is_deeply($n->subnode_ids, [9, undef, undef, undef], 'subnode ids okay (manual insertion)');
+
+=pod
 
 my $node = Algorithm::SpatialIndex::Node->new;
 my $id = $storage->store_node($node);
